moving content of generator_utils to subpixel_refinement_utils
